backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic-settings
openai
requests
beautifulsoup4
passlib[bcrypt]
httpx
python-jose